separated components for readability and better orginization, fixed modal compoenent to render avatars in free vs locked separately, fixed username bug, fixed colorpicker wheel and updated styling to better fit everything in modal
