+++ screens/profile2/components/CombinedModal/utils.js
@@ -1,0 +1,44 @@
+// utils.js
+export function hsvToRgb(h, s, v) {
+  let r, g, b;
+
+  let i = Math.floor(h * 6);
+  let f = h * 6 - i;
+  let p = v * (1 - s);
+  let q = v * (1 - f * s);
+  let t = v * (1 - (1 - f) * s);
+
+  switch (i % 6) {
+    case 0:
+      (r = v), (g = t), (b = p);
+      break;
+    case 1:
+      (r = q), (g = v), (b = p);
+      break;
+    case 2:
+      (r = p), (g = v), (b = t);
+      break;
+    case 3:
+      (r = p), (g = q), (b = v);
+      break;
+    case 4:
+      (r = t), (g = p), (b = v);
+      break;
+    case 5:
+      (r = v), (g = p), (b = q);
+      break;
+  }
+
+  return `rgb(${Math.round(r * 255)}, ${Math.round(g * 255)}, ${Math.round(
+    b * 255
+  )})`;
+}
+
+export function hsvToHex(h, s, v) {
+  const rgb = hsvToRgb(h, s, v);
+  const result = rgb
+    .match(/\d+/g)
+    .map((x) => parseInt(x).toString(16).padStart(2, "0"))
+    .join("");
+  return `#${result}`;
+}
